// src/modules/core/domain/helpers/enum/collection.enum.ts
export enum CollectionEnum {
  NAME_COLLECTION = 'core.folios.reporte',
  CORE_FOLIOS = 'core.folios',
  /** collection: core.titulares*/
  TITULARES = 'core.titulares',
  /** collection: core.polizas*/
  POLIZAS = 'core.polizas',
  /** collection: cat.unidades*/
  UNIDADES = 'cat.unidades',
  /** collection: cat.riesgos*/
  RIESGOS = 'cat.riesgos',
  /** collection: cat.tipos-movimiento*/
  TIPOS_MOVIMIENTO = 'cat.tipos-movimiento',
  /** collection: adm.proyectos*/
  PROYECTOS = 'adm.proyectos',
  /** collection: adm.aseguradoras*/
  ASEGURADORAS = 'adm.aseguradoras',
  /** collection: adm.usuarios*/
  USUARIOS = 'adm.usuarios',
  /** collection: core.informacion-ejecutivo*/
  INFORMACION_EJECUTIVO = 'core.informacion-ejecutivo',
  /** collection: core.comentarios*/
  COMENTARIOS = 'core.comentarios',
  /** collection: bitacora.actividades*/
  BITACORA_ACTIVIDADES = 'bitacora.actividades',
  /** collection: flow.contacto-telefonico*/
  FLOW_CONTACTO_TELEFONICO = 'flow.contacto-telefonico',
  /** collection: task.actividad-reminder*/
  TASK_ACTIVIDAD_REMINDER = 'task.actividad-reminder',
  /** collection: adm.configuracion-documental*/
  CONFIGURACION_DOCUMENTAL = 'adm.configuracion-documental',
  /** collection: adm.documentos*/
  DOCUMENTOS = 'adm.documentos',
  /** collection: expediente.archivos*/
  EXPEDIENTE_ARCHIVOS = 'expediente.archivos',
  /** collection: workflow.actividades*/
  WORKFLOW_ACTIVIDADES = 'workflow.actividades',
  /** collection: cat.actividades*/
  ACTIVIDADES = 'cat.actividades',
  /** collection: cat.estatus-actividades*/
  ESTATUS_ACTIVIDADES = 'cat.estatus-actividades',

  FLOW_VALIDACION_DOCUMENTAL = 'flow.validacion-documental',
}